simple ask processor implementation

+++ src/main/java/io/tbd/tbdex/protocol/messages/Ask.java
@@ -2,13 +2,14 @@ package io.tbd.tbdex.protocol.messages;
 
 import io.tbd.tbdex.protocol.core.MessageBody;
 import io.tbd.tbdex.protocol.core.MessageType;
+import java.math.BigDecimal;
 
 public class Ask extends MessageBody {
     public String sourceCurrency;
-    public int sourceAmount;
+    public BigDecimal sourceAmount;
     public String targetCurrency;
 
-    public Ask(String sourceCurrency, int sourceAmount, String targetCurrency) {
+    public Ask(String sourceCurrency, BigDecimal sourceAmount, String targetCurrency) {
         super(MessageType.Ask);
 
         this.sourceCurrency = sourceCurrency;
